Make the Options component redraw on each update
index.js:42: syncOptions.get() receives an anonymous wrapping callback instead of `renderOptions`. This is done to made differences between its behavior and `.subscribe()` behavior cleaner. `.get()` calls a callback in a Node style, whereas `.subscribe()` saves the callback and doesn’t call it in the Node style.

A new `subscribe()` method is added to syncOptions. It receives a callback or an array of callbacks and calls them each time the options are updated.